+++ src/containers/FeedPage/index.js
@@ -5,10 +5,10 @@ import { routes } from "../Router";
 import { createPost, getAllPosts } from "../../actions/post";
 import styled from "styled-components";
 import TextField from "@material-ui/core/TextField";
-import Post from "../../components/Post";
 import Button from "@material-ui/core/Button";
 import { createMuiTheme, makeStyles, ThemeProvider } from '@material-ui/core/styles';
 import LoadingRing from "../../components/LoadingRing";
+import PostList from "../PostList/PostList";
 
 class FeedPage extends Component {
   constructor(props) {
@@ -16,6 +16,7 @@ class FeedPage extends Component {
     this.state = {
       text: "",
       title: "",
+      searchInput: '',
     };
   }
 
@@ -43,7 +44,7 @@ class FeedPage extends Component {
   }
 
   render() {
-    const { text } = this.state
+    const { text, searchInput } = this.state
     const { allPosts } = this.props
 
     const theme = createMuiTheme({
@@ -82,13 +83,14 @@ class FeedPage extends Component {
             </Button>
           </PostWrapper>
           {allPosts ?
-            <PostList>
-              <Post/>
-            </PostList>
+            <PostListWrapper>
+              <PostList/>
+            </PostListWrapper>
             :
-            <PostList>
+            <PostListWrapper>
               <LoadingRing/>
-            </PostList>
+              <p>Aguarda um Cadim!</p>
+            </PostListWrapper>
           }
 
       </FeedPageWrapper>
@@ -121,6 +123,7 @@ const FeedPageWrapper = styled.div`
   height: auto;
   min-height: 75vh;
 `
+
 const PostWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -131,7 +134,7 @@ const PostWrapper = styled.div`
   height: auto;
 `
 
-const PostList = styled.div`
+const PostListWrapper = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
